Outsource the resources filter function to seperate file
